+++ product-sdk/packages/signer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @parity/product-sdk-signer
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [1cc3790]
+  - @parity/product-sdk-host@0.2.2
+  - @parity/product-sdk-keys@0.2.2
+
 ## 0.2.2
 
 ### Patch Changes