+++ product-sdk/packages/storage/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @parity/product-sdk-storage
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [1cc3790]
+  - @parity/product-sdk-host@0.2.2
+
 ## 0.1.2
 
 ### Patch Changes