+++ product-sdk/packages/tx/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @parity/product-sdk-tx
 
+## 0.2.2
+
+### Patch Changes
+
+- @parity/product-sdk-keys@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes