+++ product-sdk/scripts/bench-consumer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @parity/product-sdk-bench-consumer
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [1cc3790]
+- Updated dependencies [1cc3790]
+  - @parity/product-sdk-contracts@0.4.0
+  - @parity/product-sdk@0.4.0
+  - @parity/product-sdk-descriptors@0.4.0
+  - @parity/product-sdk-chain-client@0.4.0
+  - @parity/product-sdk-bulletin@0.4.0
+  - @parity/product-sdk-host@0.2.2
+  - @parity/product-sdk-signer@0.2.3
+  - @parity/product-sdk-statement-store@0.2.2
+  - @parity/product-sdk-storage@0.1.3
+  - @parity/product-sdk-keys@0.2.2
+  - @parity/product-sdk-tx@0.2.2
+
 ## 0.0.3
 
 ### Patch Changes